+++ src/Handlers/RankHandler.php
@@ -2,7 +2,7 @@
 
 namespace Rubix\Server\Handlers;
 
-use Rubix\ML\AnomalyDetectors\Ranking;
+use Rubix\ML\Ranking;
 use Rubix\Server\Commands\Rank;
 use Rubix\Server\Responses\RankResponse;
 
@@ -11,12 +11,12 @@ class RankHandler implements Handler
     /**
      * The ranking model that is being served.
      *
-     * @var \Rubix\ML\AnomalyDetectors\Ranking
+     * @var \Rubix\ML\Ranking
      */
     protected $estimator;
 
     /**
-     * @param \Rubix\ML\AnomalyDetectors\Ranking $estimator
+     * @param \Rubix\ML\Ranking $estimator
      */
     public function __construct(Ranking $estimator)
     {