+++ src/RPCServer.php
@@ -3,16 +3,18 @@
 namespace Rubix\Server;
 
 use Rubix\ML\Learner;
+use Rubix\ML\Ranking;
 use Rubix\ML\Estimator;
 use Rubix\ML\Probabilistic;
-use Rubix\ML\AnomalyDetectors\Ranking;
 use Rubix\Server\Commands\Command;
 use Rubix\Server\Commands\Predict;
+use Rubix\Server\Commands\PredictSample;
 use Rubix\Server\Commands\Proba;
 use Rubix\Server\Commands\Rank;
 use Rubix\Server\Commands\QueryModel;
 use Rubix\Server\Commands\ServerStatus;
 use Rubix\Server\Handlers\PredictHandler;
+use Rubix\Server\Handlers\PredictSampleHandler;
 use Rubix\Server\Handlers\ProbaHandler;
 use Rubix\Server\Handlers\RankHandler;
 use Rubix\Server\Handlers\QueryModelHandler;
@@ -241,6 +243,10 @@ class RPCServer implements Server, LoggerAware
             ServerStatus::class => new ServerStatusHandler($this),
             Predict::class => new PredictHandler($estimator),
         ];
+
+        if ($estimator instanceof Learner) {
+            $commands[PredictSample::class] = new PredictSampleHandler($estimator);
+        }
 
         if ($estimator instanceof Probabilistic) {
             $commands[Proba::class] = new ProbaHandler($estimator);